Add a Polynomial Curve Fitter
Since starting at an ML algorithm out of nowhere would be ludicrous, a curve fitter was the logical first step for me. This programming exercise acted as an oversimplified version of a neural network, and as you will see, some parts of the code were reused in the ML algorithm (A neural network is basically a multi-dimensional curve fitter).